removed white "halo" surrounding Textures
when creating tilemaps, the rendered side-by-side textures destroys the "map illusion" because of the grid created by the white-ish halo surrounding each texture.  this is a quick and easy fix but somehow, the white halo has just become black-ish. but it works on my end so, committing the changes now.